fix issues

+++ src/main/scala/io/picnicml/examples/sarcasm/TrainClassifier.scala
@@ -37,7 +37,7 @@ object TrainClassifier extends App {
   def shuffleSplitData(x: Features, y: Target): TrainTestSplit = {
     println("Shuffling and splitting data")
     val (xShuffled, yShuffled) = shuffleDataset(x, y)
-    val split = splitDataset(xShuffled, yShuffled, proportionTrain = 0.9)
+    val split = splitDataset(xShuffled, yShuffled, proportionTrain = 0.9f)
     println(s"Training set size: ${split.xTr.rows}, test set size: ${split.xTe.rows}")
     println(s"Proportion of sarcastic examples in training set: ${sum(split.yTr) / split.xTr.rows}")
     println(s"Proportion of sarcastic examples in test set: ${sum(split.yTe) / split.xTe.rows}\n")
@@ -49,12 +49,12 @@ object TrainClassifier extends App {
     val cv = CrossValidation(f1Score, KFoldSplitter(numFolds = 10))
     val search = HyperparameterSearch(numGridSearchIterations, cv)
     val (start, end, step) = (1e-5, 5.0, (5.0 - 1e-5) / numGridSearchIterations)
-    val grid = Range.BigDecimal(start, end, step).map(_.toDouble).iterator
+    val grid = Range.BigDecimal(start, end, step).map(_.toFloat).iterator
     println("Searching the hyperparameter space")
     search.bestOf(split.xTr, split.yTr) { generateModel(lambda = grid.next) }
   }
 
   // lambda is L2 regularization strength
-  def generateModel(lambda: Double): Pipeline =
+  def generateModel(lambda: Float): Pipeline =
     Pipeline(List(pipe(StandardScaler(featureIndex))))(pipe(LogisticRegression(lambda)))
 }
